refactor: Refactor packages
