--- application.pid
@@ -1,1 +1,0 @@
-19488
