+++ src/main/java/org/isf/usergroups/dto/UserGroupDTO.java
@@ -19,12 +19,14 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package org.isf.menu.dto;
+package org.isf.usergroups.dto;
 
 import java.util.List;
+
 import jakarta.validation.constraints.NotNull;
 
 import org.isf.permissions.dto.PermissionDTO;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 
 public class UserGroupDTO {
