+++ src/main/java/org/isf/usergroups/mapper/UserGroupMapper.java
@@ -19,15 +19,16 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package org.isf.menu.mapper;
+package org.isf.usergroups.mapper;
 
-import org.isf.menu.dto.UserGroupDTO;
+import org.isf.usergroups.dto.UserGroupDTO;
 import org.isf.menu.model.UserGroup;
 import org.isf.shared.GenericMapper;
 import org.springframework.stereotype.Component;
 
 @Component
 public class UserGroupMapper extends GenericMapper<UserGroup, UserGroupDTO> {
+
 	public UserGroupMapper() {
 		super(UserGroup.class, UserGroupDTO.class);
 	}
