+++ src/main/java/org/isf/users/dto/UserDTO.java
@@ -19,7 +19,9 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package org.isf.menu.dto;
+package org.isf.users.dto;
+
+import org.isf.usergroups.dto.UserGroupDTO;
 
 import jakarta.validation.constraints.NotNull;
