+++ src/main/java/org/isf/usersettings/mapper/UserSettingMapper.java
@@ -19,15 +19,16 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package org.isf.menu.mapper;
+package org.isf.usersettings.mapper;
 
-import org.isf.menu.dto.UserSettingDTO;
+import org.isf.usersettings.dto.UserSettingDTO;
 import org.isf.menu.model.UserSetting;
 import org.isf.shared.GenericMapper;
 import org.springframework.stereotype.Component;
 
 @Component
 public class UserSettingMapper extends GenericMapper<UserSetting, UserSettingDTO> {
+
 	public UserSettingMapper() {
 		super(UserSetting.class, UserSettingDTO.class);
 	}
