+++ src/test/java/org/isf/users/rest/UserControllerTest.java
@@ -27,7 +27,9 @@ import java.util.List;
 
 import org.isf.OpenHospitalApiApplication;
 import org.isf.menu.manager.UserBrowsingManager;
-import org.isf.menu.mapper.UserMapper;
+import org.isf.users.data.UserHelper;
+import org.isf.users.dto.UserDTO;
+import org.isf.users.mapper.UserMapper;
 import org.isf.menu.model.User;
 import org.isf.menu.model.UserGroup;
 import org.isf.permissions.manager.PermissionManager;
@@ -39,14 +41,20 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+import static org.hamcrest.Matchers.containsString;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.log;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import org.slf4j.Logger;
@@ -80,6 +88,113 @@ public class UserControllerTest {
 
 	@MockBean
 	private PermissionManager permissionManager;
+
+	@Nested
+	@DisplayName("Get users")
+	class GetUsers {
+
+		@Test
+		@WithMockUser(username = "admin", authorities = { "users.read" })
+		@DisplayName("Get all users")
+		void getAllUsers() throws Exception {
+			List<User> users = UserHelper.generateUsers(3);
+			List<UserDTO> userDTOS = userMapper.map2DTOList(users);
+
+			when(userManager.getUser()).thenReturn(users);
+
+			var result = mvc.perform(
+					get("/users").contentType(MediaType.APPLICATION_JSON))
+				.andDo(log())
+				.andExpect(status().isOk())
+				.andExpect(content().string(containsString(UserHelper.asJsonString(userDTOS.stream().peek(user -> user.setPasswd(null)).toList()))))
+				.andReturn();
+
+			LOGGER.debug("result: {}", result);
+		}
+
+		@Test
+		@WithMockUser(username = "admin", authorities = { "users.read" })
+		@DisplayName("Get users by group code")
+		void getAllUsersByGroupCode() throws Exception {
+			List<User> users = UserHelper.generateUsers(3);
+			List<UserDTO> userDTOS = userMapper.map2DTOList(users);
+			String groupCode = "contrib";
+
+			when(userManager.getUser(any())).thenReturn(users);
+
+			var result = mvc.perform(
+					get("/users").queryParam("group_id", groupCode).contentType(MediaType.APPLICATION_JSON))
+				.andDo(log())
+				.andExpect(status().isOk())
+				.andExpect(content().string(containsString(UserHelper.asJsonString(userDTOS.stream().peek(user -> user.setPasswd(null)).toList()))))
+				.andReturn();
+
+			LOGGER.debug("result: {}", result);
+		}
+
+		@Test
+		@WithMockUser(username = "admin", authorities = { "users.read" })
+		@DisplayName("Get user by name")
+		void getUserByName() throws Exception {
+			User user = UserHelper.generateUser();
+			UserDTO userDTO = userMapper.map2DTO(user);
+			String username = "admin";
+
+			when(userManager.getUserByName(any())).thenReturn(user);
+
+			userDTO.setPasswd(null);
+
+			var result = mvc.perform(
+					get("/users/{username}", username).contentType(MediaType.APPLICATION_JSON))
+				.andDo(log())
+				.andExpect(status().isOk())
+				.andExpect(content().string(containsString(UserHelper.asJsonString(userDTO))))
+				.andReturn();
+
+			LOGGER.debug("result: {}", result);
+		}
+	}
+
+	@Test
+	@WithMockUser(username = "admin", authorities = { "users.create" })
+	@DisplayName("Create user")
+	void createUser() throws Exception {
+		User user = UserHelper.generateUser();
+		UserDTO userDTO = userMapper.map2DTO(user);
+
+		when(userManager.newUser(any())).thenReturn(user);
+
+		var result = mvc.perform(
+				post("/users")
+					.contentType(MediaType.APPLICATION_JSON)
+					.content(objectMapper.writeValueAsString(userDTO))
+			)
+			.andDo(log())
+			.andExpect(status().isCreated())
+			.andExpect(content().string(containsString("true")))
+			.andReturn();
+
+		LOGGER.debug("result: {}", result);
+	}
+
+	@Test
+	@WithMockUser(username = "admin", authorities = { "users.delete" })
+	@DisplayName("Delete user")
+	void deleteUser() throws Exception {
+		User user = UserHelper.generateUser();
+
+		when(userManager.getUserByName(any())).thenReturn(user);
+		doNothing().when(userManager).deleteUser(any());
+
+		var result = mvc.perform(
+				delete("/users/{username}", user.getUserName()).contentType(MediaType.APPLICATION_JSON))
+			.andDo(log())
+			.andExpect(status().isOk())
+			.andExpect(content().string(containsString("true")))
+			.andReturn();
+
+		LOGGER.debug("result: {}", result);
+	}
 
 	@Nested
 	@DisplayName("Update user")
